requirements.txt
pyserial~=3.5
pytest~=8.1.1
python-dotenv~=1.0.1
fastapi~=0.110.1
uvicorn[standard]
numpy
